Comando IN no MySQL

+++ index.php
@@ -7,48 +7,80 @@ include_once "conexao.php";
 
 <head>
     <meta charset="UTF-8">
-    <title>LIKE</title>
+    <title>IN</title>
     <link rel="shortcut icon" href="images/favicon.ico">
 </head>
 
 <body>
     <h2>Listar usuários</h2>
+
     <?php
-    //recebe os dados do formulário
     $dados = filter_input_array(INPUT_POST, FILTER_DEFAULT);
     ?>
 
     <form method="POST" action="">
-        <?php
-        $result_nome = "";
-            if(isset($dados['texto_pesquisar'])){
-                $result_nome = $dados['texto_pesquisar'];
-            }
-        ?>
-        <label>Pesquisar</label>
-        <input type="text" name="texto_pesquisar" placeholder="Pesquisar pelo termo" value="<?php echo $result_nome;?>" /><br><br>
+        <label>Pesquisar</label><br><br>
 
-        <input type="submit" value="Pesquisar" name="PesqUsuario"><br><br>
+        <?php
+        //Pesquisa os niveis de acesso no BD
+        $query_niveis_acessos = "SELECT id, nome FROM niveis_acessos ORDER BY ordem ASC";
+        $result_niveis_acessos = $conn->prepare($query_niveis_acessos);
+        $result_niveis_acessos->execute();
+
+        $valor_pesq_list = "";
+        if (!empty($dados['nivel_acesso'])) {
+            foreach ($dados['nivel_acesso'] as $niveis_acesso_id) {
+                $valor_pesq_list .= "$niveis_acesso_id, ";
+            }
+        }
+
+        while ($row_nivel_acesso = $result_niveis_acessos->fetch(PDO::FETCH_ASSOC)) {
+            extract($row_nivel_acesso);
+            $result_valor = mb_strpos($valor_pesq_list, $id);
+            if ($result_valor === false) {
+                $checked = "";
+            } else {
+                $checked = "checked";
+            }
+
+            echo "<input type='checkbox' name='nivel_acesso[]' value='$id' $checked>$nome<br>";
+        }
+        ?>
+
+        <br><input type="submit" value="Pesquisar" name="PesqUsuario"><br><br>
+
     </form>
 
     <?php
     if (!empty($dados['PesqUsuario'])) {
-        //% - indica que aceita texto antes ou depois
-        $nome = "%" . $dados['texto_pesquisar'] . "%";
-        $query_usuario = "SELECT id, nome, email FROM usuarios WHERE nome LIKE :nome ORDER BY id DESC";
-        $result_usuarios = $conn->prepare($query_usuario);
-        $result_usuarios->bindParam(':nome', $nome, PDO::PARAM_STR);
-        $result_usuarios->execute();
 
-        while ($row_usuario = $result_usuarios->fetch(PDO::FETCH_ASSOC)) {
-            extract($row_usuario);
-            echo "ID: $id <br>";
-            echo "Nome: $nome <br>";
-            echo "E-mail: $email <br>";
-            echo "<hr>";
+        $valor_pesq = "";
+        $controle = 1;
+        if (array_key_exists('nivel_acesso', $dados)) {
+
+            foreach ($dados['nivel_acesso'] as $niveis_acesso_id) {
+                if ($controle == 1) {
+                    $valor_pesq = $niveis_acesso_id;
+                } else {
+                    $valor_pesq .= ", $niveis_acesso_id";
+                }
+                $controle++;
+            }
+            $query_usuarios = "SELECT id, nome, email, niveis_acesso_id FROM usuarios WHERE niveis_acesso_id IN ($valor_pesq) ORDER BY id DESC";
+            $result_usuarios = $conn->prepare($query_usuarios);
+            $result_usuarios->execute();
+
+            while ($row_usuario = $result_usuarios->fetch(PDO::FETCH_ASSOC)) {
+                extract($row_usuario);
+
+                echo "ID: $id <br>";
+                echo "Nome: $nome <br>";
+                echo "E-mail: $email <br>";
+                echo "Nível de acesso: $niveis_acesso_id <br>";
+                echo "<hr>";
+            }
         }
     }
-
     ?>
 </body>
